s/pos/(unsigned) pos/
svn path=/trunk/kdelibs/; revision=85944

+++ khtml/css/css_valueimpl.cpp
@@ -169,7 +169,7 @@ void CSSStyleDeclarationImpl::setProperty(int id, const DOMString &value, bool i
 	    p->nonCSSHint = true;
 	    p = m_lstValues->next();
 	}
-    } else if(pos == m_lstValues->count() )
+    } else if((unsigned) pos == m_lstValues->count() )
 	kdDebug( 6080 ) << "CSSStyleDeclarationImpl::setProperty invalid property=" << id << "value: " << value.string() << endl;
     if (m_node)
 	m_node->setChanged(true);
